added notebooks
